Refactor code

+++ x-pack/plugin/security/src/internalClusterTest/java/org/elasticsearch/test/NativeRealmIntegTestCase.java
@@ -43,9 +43,15 @@ public abstract class NativeRealmIntegTestCase extends SecurityIntegTestCase {
         }
     }
 
+    public boolean shouldDeleteSecurityIndex() {
+        return true;
+    }
+
     @After
     public void stopESNativeStores() throws Exception {
-        deleteSecurityIndex();
+        if (shouldDeleteSecurityIndex()) {
+            deleteSecurityIndex();
+        }
 
         if (getCurrentClusterScope() == Scope.SUITE) {
             // Clear the realm cache for all realms since we use a SUITE scoped cluster
